game hub pass message to game loader

+++ internal/gamemanager/message/unit_move.go
@@ -2,8 +2,8 @@ package message
 
 // UnitMoveMessageData is message data for CmdUnitMove
 type UnitMoveMessageData struct {
-	Y1 uint8 `json:"y_1"`
-	X1 uint8 `json:"x_1"`
-	Y2 uint8 `json:"y_2"`
-	X2 uint8 `json:"x_2"`
+	Y1 int `json:"y_1"`
+	X1 int `json:"x_1"`
+	Y2 int `json:"y_2"`
+	X2 int `json:"x_2"`
 }
